pyproject.toml
[project]
name = "pciv"
version = "0.1.0"
description = "Plan-Critique-Implement-Verify multi-agent orchestration CLI."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "pciv contributors" }]

dependencies = [
    "agent-framework>=1.0.0b1",
    "openai>=1.54.0",
    "pydantic>=2.9.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.2",
    "opentelemetry-sdk>=1.30.0",
    "azure-monitor-opentelemetry>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
pciv = "pciv.cli:app"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pciv"]

[tool.hatch.build.targets.wheel.force-include]
"src/pciv/state/schema.sql" = "pciv/state/schema.sql"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"
packages = ["pciv"]

[[tool.mypy.overrides]]
module = [
    "agent_framework.*",
    "azure.monitor.opentelemetry.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"
